Final Arcade teleop

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/nullTeleopMac.java
@@ -18,7 +18,7 @@ public class nullTeleopMac extends OpMode {
     DcMotor motorRB, motorRF, motorLB, motorLF;
     //spinner, shooter;
     int threshold = 10;
-    double rightY, leftY, rightX, leftX;
+    float rightY, leftY, rightX, leftX;
    // boolean spin, shoot;
 
     public nullTeleopMac() {
